requirements.txt
-i https://pypi.org/simple

bcrypt==4.0.1

cffi==1.15.1

click==8.1.3

colorama==0.4.6

cryptography==39.0.2

feedparser==6.0.10

flask==2.2.3

flask-bcrypt==1.0.1

flask-cors==3.0.10

gunicorn==20.1.0

itsdangerous==2.1.2

jinja2==3.1.2

markupsafe==2.1.2

pycparser==2.21

pymysql==1.0.2

pyopenssl==23.0.0

python-dotenv==1.0.0

setuptools==67.6.1; python_version >= '3.7'

sgmllib3k==1.0.0

six==1.16.0

waitress==2.1.2

werkzeug==2.2.3

